extract routes table components
