+++ src/components/routes-table/index.js
@@ -1,0 +1,1 @@
+export * from './routes-table'
